+++ CHANGELOG.md
@@ -17,7 +17,9 @@
 - new Dart code file structure #96 
 - Base the sdk name on the platform (`sentry.dart` for io & flutter, `sentry.dart.browser` in a browser context) #103 
 - Single changelog and readme for both packages #105
+- new static API : Sentry.init(), Sentry.captureEvent() #108
 - expect a sdkName based on the test platform #105
+- Added Scope and Breadcrumb ring buffer #109
 
 # `package:sentry` changelog